separate out query module, split query apply into apply and parse
